pyproject.toml
[project]
name = "devresources-django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
]
